Wasm funcs are moved to core
